Fix direnv on Vim 8.0
With the version of Vim 8.0 I have (MacVim, 8.0, patches 1-1272), I am getting
this error with direnv.vim:

  Error detected while processing function
  .../direnv.vim/plugin/direnv.vim|60| direnv#export[14]
  E475: Invalid argument: stderr

It appears that in Vim 8.0, the dictionary provided to the job must not have
any keys *except* the callback keys, unlike with neovim.

After several experimental implementations, this is the one that works.