Fix direnv.vim blocking the editor when the direnv binary is missing
direnv.vim currently calls :echoerr to notify users when the direnv
binary isn't found in PATH. However, this causes vim to output 3 lines
of messages which results in the hit-enter prompt to be shown, blocking
the editor. This warning message can be shown frequently throughout an
editing session, causing annoyances for users working on a system
without the direnv binary present. This change replaces :echoerr with
:echom to reduce the number of lines output to 1. This should prevent
direnv.vim from blocking the editor in most configurations.